pyproject.toml

[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = [{name = "Your Name", email = "you@example.com"}]
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.1",
    "schedule>=1.2.2",
    "python-telegram-bot[job-queue,webhooks]>=20.0",
    "telegram>=0.0.1",
    "telethon>=1.25.0",
    "langgraph>=0.3.5",
    "langchain-anthropic>=0.3.9",
    "langchain-core>=0.3.43",
]
